Update ConvertPDFtoXLSX_tabula_v2.py
The only issue with using tkinter is that the window stays open permanently. I haven't managed to find a solution to close it at the end of the program without getting an error. Also need to work out how to open file location and type a file name without having to hard code it in.